remove composite
