+++ lib/subscribe/parent/origin.js
@@ -1,7 +1,6 @@
 'use strict'
 const diff = require('../diff')
 const remove = require('../remove')
-// const composite = require('./composite')
 
 module.exports = function parent (target, subs, update, tree, stamp) {
   var parentTree = tree.parent
