+++ lib/subscribe/parent/subscription.js
@@ -1,7 +1,6 @@
 'use strict'
 const diff = require('../diff')
 const remove = require('../remove')
-const composite = require('./composite')
 
 module.exports = function parent (target, subs, update, tree, stamp) {
   var parentTree = tree.$parent
@@ -38,4 +37,25 @@ function get (target, tree) {
   }
   // can be greatly optmized -- especially root walking all the time
   return (!path.length) ? target.getRoot() : target.getRoot().get(path)
+}
+
+function composite (tree, type) {
+  var key = type
+  var parentcounter = 1
+  while (tree._p && parentcounter) {
+    if (tree._key !== type) {
+      // add both! -- difficult to combine
+      // have to do
+      if (!tree.$c) { tree.$c = {} }
+      tree.$c[key] = type
+      key = tree._key
+      tree = tree._p
+      if (key !== '$any') {
+        parentcounter--
+      }
+    } else {
+      parentcounter++
+      tree = tree._p
+    }
+  }
 }
